requirements.txt
parakeet-mlx>=0.5
sounddevice>=0.5
rumps>=0.4
pynput>=1.8
pyobjc-framework-Quartz>=12.0
pyobjc-framework-Cocoa>=12.0
pyobjc-framework-ApplicationServices>=12.0
numpy
soundfile
